Some variable types changed.
<duc>, <wuc> and <tuc> variable types changed as "single".

 Changes to be committed:
	modified:   assignAppliances2endUsers.m
	modified:   checkAlwaysNever.m
	modified:   worktime_nonPeriodic.m
	modified:   worktime_periodic_nonContinous.m


Former-commit-id: a5c08ff746b21a0a68568ea266f7d4fb9d24c212

+++ lib/dataGenerator/assignAppliances2endUsers.m
@@ -102,10 +102,10 @@ function endUsers = assignAppliances2endUsers(appliancesData, electricVehicles, 
 				end
 
 				if ~(continuity)
-					endUsers(i).appliances.(string(independentAppliances(index))).duc = uint16(0);
-					endUsers(i).appliances.(string(independentAppliances(index))).wuc = uint16(0);
+					endUsers(i).appliances.(string(independentAppliances(index))).duc = single(0);
+					endUsers(i).appliances.(string(independentAppliances(index))).wuc = single(0);
 				end
-				endUsers(i).appliances.(string(independentAppliances(index))).tuc = uint16(0);
+				endUsers(i).appliances.(string(independentAppliances(index))).tuc = single(0);
 			end
 		end
 		
@@ -130,10 +130,10 @@ function endUsers = assignAppliances2endUsers(appliancesData, electricVehicles, 
 					end
 
 					if ~(continuity)
-						endUsers(i).appliances.(string(dependentAppliances(index))).duc = uint16(0);
-						endUsers(i).appliances.(string(dependentAppliances(index))).wuc = uint16(0);
+						endUsers(i).appliances.(string(dependentAppliances(index))).duc = single(0);
+						endUsers(i).appliances.(string(dependentAppliances(index))).wuc = single(0);
 					end
-					endUsers(i).appliances.(string(dependentAppliances(index))).tuc = uint16(0);
+					endUsers(i).appliances.(string(dependentAppliances(index))).tuc = single(0);
 				end
 			end
 		end
